requirements.txt
torch==2.0.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
transformers==4.30.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
umap-learn==0.5.3
shap==0.42.1
ipykernel==6.25.0
jupyterlab==4.0.5
nbformat==5.9.2
requests==2.31.0
